+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### dependabot: \#116 Bump next from 13.0.2 to 14.2.10 in /cypress
+
 ### dependabot: \#114 Bump the github-actions group with 3 updates
 
 ## [2.13.0] - 2024-10-16