+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.13.1] - 2024-12-02
+
 ### Fixed
 
 - pick disabled options on Tab, updating react-bootstrap-typeahead to v6.3.4.
@@ -429,7 +431,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.13.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.13.1...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -454,6 +456,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[2.13.1]: https://github.com/neolution-ch/react-hook-form-components/compare/2.13.0...2.13.1
 [2.13.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.12.0...2.13.0
 [2.12.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.11.0...2.12.0
 [2.11.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.10.0...2.11.0